+++ lib/honeybadger.rb
@@ -14,6 +14,7 @@ require 'honeybadger/stats'
 require 'honeybadger/user_informer'
 require 'honeybadger/user_feedback'
 require 'honeybadger/integrations'
+require 'honeybadger/exception_extensions'
 
 require 'honeybadger/railtie' if defined?(Rails::Railtie)